Move functions to graph.ts
