+++ optuna_dashboard/ts/components/GraphContour.tsx
@@ -22,7 +22,7 @@ import * as plotly from "plotly.js-dist-min"
 import React, { FC, useEffect, useMemo, useState } from "react"
 import { SearchSpaceItem, StudyDetail, Trial } from "ts/types/optuna"
 import { PlotType } from "../apiClient"
-import { getAxisInfo } from "../graphUtil"
+import { getAxisInfo } from "../graph"
 import { usePlot } from "../hooks/usePlot"
 import { usePlotlyColorTheme } from "../state"
 import { useBackendRender } from "../state"
