+++ tslib/react/src/index.ts
@@ -23,6 +23,6 @@ export {
   getIsDominated,
 } from "./utils/trialFilter"
 export { useMergedUnionSearchSpace } from "./utils/searchSpace"
-export { makeHovertext, getAxisInfo } from "./utils/graphUtil"
-export type { AxisInfo } from "./utils/graphUtil"
+export { makeHovertext, getAxisInfo } from "./utils/graph"
+export type { AxisInfo } from "./utils/graph"
 export type { GraphComponentState } from "./types"
